refactor: rename use-contact-form and update import path in Contact.tsx
